requirements.txt
asgiref==3.3.4
boto3==1.17.94
botocore==1.20.94
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-storages==1.11.1
expertai-nlapi==2.3.1
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
nltk==3.6.2
numpy==1.20.3
oauthlib==3.1.1
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
textblob==0.15.3
tqdm==4.61.1
tweepy==3.10.0
urllib3==1.26.5
